requirements.txt
# flask==2.3.3
# python-dotenv==1.0.0
# supabase==2.0.3
# numpy==1.24.3
# opencv-python==4.8.0.76  # Sửa từ 4.8.0 thành 4.8.0.76
# moviepy==1.0.3
# tensorflow==2.15.0
# pillow==10.0.0
# python-magic==0.4.27
# tqdm==4.66.1
# scikit-learn==1.3.0


python-dotenv==1.0.0
supabase==2.3.0
opencv-python==4.8.1.78
numpy==1.26.0
scikit-learn==1.3.2
pillow==10.1.0
flask==3.0.0
matplotlib==3.8.0
tensorflow==2.15.0  # For object detection